flask-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.15.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
